Use a class to pass progress info back to main thread
This was done with a dict with string typing and some
kubernetes specific naming (like 'pod') now. This commit
instead makes use of a class, with enums for the different
possible payloads. Most 'magic' strings are removed, and
it is clearer to see when all cases that need to be handled
are handled.

During this, I discovered that we had no special handling for
when a pod enters Failed state. I've explicitly included the
state, but preserved the current behavior.

Add a little bit of typing information in there while
at it, to make life easier in the future.