docs: Fix formatting and remove outdated advice
Note that the alignment advice is problematic. Not each line needs to be
left aligned, but it is true that no line should go further left than
the initial line. The part about not forgetting | becomes a bit
redundant as we have schema validation now that will make sure it is set
to be a string.